--- legend-engine-server-shared/src/test/resources/config-test.yaml
@@ -1,50 +1,0 @@
-# Copyright 2020 Goldman Sachs
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-applicationName: Legend Data Push Test Server
-
-sessionCookie: LEGEND_SDLC_JSESSIONID
-
-server:
-  applicationConnectors:
-    - type: http
-      port: 8888
-      maxRequestHeaderSize: 128KiB
-  adminConnectors:
-    - type: http
-      port: 8889
-  gzip:
-    includedMethods:
-      - GET
-      - POST
-  rootPath: /api
-
-filterPriorities:
-  org.pac4j.j2e.filter.CallbackFilter: 2
-  org.pac4j.j2e.filter.SecurityFilter: 3
-  CORS: 4
-
-
-logging:
-  # Change this to affect library class logging
-  level: OFF
-  appenders:
-    - type: console
-      logFormat: "%d{yyyy-MM-dd HH:mm:ss.SSS} %-5p [%thread] %c - %m%n"
-
-swagger:
-  resourcePackage: org.finos.legend.engine.datapush.server.resources
-  title: Legend Data Push
-  version: local-snapshot
-  schemes: []
