+++ legend-engine-server-support/legend-engine-server-support-core/pom.xml
@@ -3,14 +3,14 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>legend-engine-xts-data-push</artifactId>
+        <artifactId>legend-engine-server-support</artifactId>
         <groupId>org.finos.legend.engine</groupId>
         <version>4.17.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>legend-engine-xt-data-push-server</artifactId>
-    <name>Legend Engine - XT - Data Push - Server</name>
+    <artifactId>legend-engine-server-support-core</artifactId>
+    <name>Legend Engine - Server Support - Core</name>
     <packaging>jar</packaging>
 
     <properties>
@@ -18,37 +18,8 @@
         <maven.compiler.target>11</maven.compiler.target>
     </properties>
 
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-dependency-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>dependency-analyze</id>
-                        <configuration>
-                            <!-- The dependency plugin analyzer raises these as false positives -->
-                            <ignoredUnusedDeclaredDependencies>
-                            </ignoredUnusedDeclaredDependencies>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
     <dependencies>
         <!-- Shared Drop Wizard Bundles -->
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-server-support-core</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.finos.legend.shared</groupId>
-                    <artifactId>legend-shared-pac4j-gitlab</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
         <dependency>
             <groupId>org.finos.legend.shared</groupId>
             <artifactId>legend-shared-server</artifactId>
@@ -60,13 +31,6 @@
             </exclusions>
         </dependency>
         <!-- Shared Drop Wizard Bundles -->
-
-        <!-- Data Push -->
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-xt-data-push-specification-model</artifactId>
-        </dependency>
-        <!-- Data Push -->
 
         <!-- JSON -->
         <!--
@@ -83,18 +47,7 @@
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.jaxrs</groupId>
-            <artifactId>jackson-jaxrs-json-provider</artifactId>
-        </dependency>
         <!-- JSON -->
-
-        <!-- APACHE HTTP CLIENT -->
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-        </dependency>
-        <!-- APACHE HTTP CLIENT -->
 
         <!-- ECLIPSE COLLECTIONS -->
         <dependency>
@@ -231,10 +184,12 @@
             <groupId>javax.ws.rs</groupId>
             <artifactId>javax.ws.rs-api</artifactId>
         </dependency>
+        <!--
         <dependency>
             <groupId>io.swagger</groupId>
             <artifactId>swagger-annotations</artifactId>
         </dependency>
+        -->
         <!-- ANNOTATIONS -->
 
         <!-- RUNTIME -->
@@ -265,25 +220,6 @@
         </dependency>
         <!-- RUNTIME -->
 
-        <!-- TEST -->
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.finos.legend.engine</groupId>
-            <artifactId>legend-engine-test-server-shared</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>io.dropwizard</groupId>
-            <artifactId>dropwizard-testing</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <!-- TEST -->
     </dependencies>
 
 </project>
