+++ legend-engine-server/pom.xml
@@ -59,7 +59,7 @@
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.finos.legend.engine.server.Server</mainClass>
+                                    <mainClass>org.finos.legend.engine.org.finos.legend.engine.server.support.server.Server</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
